factor out coverage code
