+++ api/main.js
@@ -1,5 +1,4 @@
 require('./env');
-const fs = require('fs');
 const logger = require('./logger')('main');
 const express = require('express');
 const cors = require('cors');
@@ -8,27 +7,11 @@ const uuid = require('uuid/v1');
 require('./db').setup();
 const auth = require('./auth');
 const routes = require('./routes');
+const endpointCoverage = require('./endpointCoverageMiddleware');
 
 const server = express();
 
-if (process.env.ENDPOINT_COVERAGE_CAPTURE) {
-  server.use((req, res, next) => {
-    const end = res.end.bind(res);
-    res.end = (...args) => {
-      const path = req.route ? req.route.path : req.path;
-      const result = {
-        path,
-        method: req.method,
-        status: res.statusCode
-      };
-      fs.appendFileSync('./endpoints-hit.txt', `${JSON.stringify(result)}\n`);
-      return end(...args);
-    };
-
-    next();
-  });
-}
-
+endpointCoverage.registerCoverageMiddleware(server);
 server.use((req, res, next) => {
   req.id = uuid();
   logger.verbose(req, `got ${req.method} request to ${req.path}`);
@@ -49,14 +32,5 @@ routes(server);
 logger.silly('starting the server');
 server.listen(process.env.PORT, () => {
   logger.verbose(`server listening on :${process.env.PORT}`);
-
-  if (process.env.ENDPOINT_COVERAGE_CAPTURE) {
-    const allRoutes = server._router.stack // eslint-disable-line no-underscore-dangle
-      .filter(a => !!a.route)
-      .map(({ route: { path, methods } }) => ({
-        path,
-        method: Object.keys(methods)[0]
-      }));
-    fs.writeFileSync('./endpoints-all.txt', JSON.stringify(allRoutes));
-  }
+  endpointCoverage.getCoverageEndpoints(server);
 });
